Introduce associated type on DocumentStore
