+++ p2panda-rs/src/test_utils/memory_store/provider.rs
@@ -9,7 +9,6 @@ use crate::hash::Hash;
 use crate::identity::PublicKey;
 use crate::operation::OperationId;
 use crate::schema::SchemaId;
-use crate::storage_provider::traits::DocumentStore;
 use crate::test_utils::memory_store::{PublishedOperation, StorageEntry};
 
 type PublickeyLogId = String;
@@ -29,5 +28,3 @@ pub struct MemoryStore {
     /// Stored operations
     pub operations: Arc<Mutex<HashMap<OperationId, PublishedOperation>>>,
 }
-
-impl DocumentStore for MemoryStore {}
