+++ p2panda-rs/src/test_utils/memory_store/stores/mod.rs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+mod document;
 mod entry;
 mod log;
 pub mod operation;
